ai-resume-analyzer/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0
openai==1.30.0
python-docx==1.1.2
PyMuPDF==1.24.4
pdfplumber==0.11.0
pydantic==2.7.0
python-dotenv==1.0.1
